+++ src/it/scala/uk/gov/hmrc/govukfrontend/views/components/GovukCharacterCountIntegrationSpec.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 HM Revenue & Customs
+ * Copyright 2019 HM Revenue & Customs
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,13 +14,15 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.govukfrontend.views
-package html
+package uk.gov.hmrc.govukfrontend.views.components
 
-package object components extends Utils with Aliases {
+import uk.gov.hmrc.govukfrontend.support.TemplateIntegrationSpec
+import uk.gov.hmrc.govukfrontend.views.html.components._
+import uk.gov.hmrc.govukfrontend.views.viewmodels.charactercount.CharacterCount
+import uk.gov.hmrc.govukfrontend.views.viewmodels.charactercount.Generators._
 
-  /**
-    * Top-level implicits for all components
-    */
-  object implicits extends Implicits
-}
+object GovukCharacterCountIntegrationSpec
+    extends TemplateIntegrationSpec[CharacterCount, GovukCharacterCount](
+      govukComponentName = "govukCharacterCount",
+      seed = None
+    )