+++ src/it/scala/uk/gov/hmrc/govukfrontend/views/components/GovukCheckboxesIntegrationSpec.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 HM Revenue & Customs
+ * Copyright 2019 HM Revenue & Customs
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,12 +15,12 @@
  */
 
 package uk.gov.hmrc.govukfrontend.views
-package html
+package components
 
-package object components extends Utils with Aliases {
+import uk.gov.hmrc.govukfrontend.support.TemplateIntegrationSpec
+import uk.gov.hmrc.govukfrontend.views.html.components._
+import uk.gov.hmrc.govukfrontend.views.viewmodels.checkboxes.Checkboxes
+import uk.gov.hmrc.govukfrontend.views.viewmodels.checkboxes.Generators._
 
-  /**
-    * Top-level implicits for all components
-    */
-  object implicits extends Implicits
-}
+object GovukCheckboxesIntegrationSpec
+    extends TemplateIntegrationSpec[Checkboxes, GovukCheckboxes](govukComponentName = "govukCheckboxes", seed = None)