+++ src/test/scala/uk/gov/hmrc/govukfrontend/views/components/GovukNotificationBannerSpec.scala
@@ -15,12 +15,9 @@
  */
 
 package uk.gov.hmrc.govukfrontend.views
-package html
+package components
 
-package object components extends Utils with Aliases {
+import uk.gov.hmrc.govukfrontend.views.html.components._
 
-  /**
-    * Top-level implicits for all components
-    */
-  object implicits extends Implicits
-}
+class GovukNotificationBannerSpec
+    extends TemplateUnitSpec[NotificationBanner, GovukNotificationBanner]("govukNotificationBanner")